Refactor Commands to handle specific commands directly
