+++ docs/README.md
@@ -1,6 +1,6 @@
-# TommyTalks User Guide
+# tommyTalks User Guide
 
-![Screenshot of TommyTalks in action.](Ui.png)
+![Screenshot of tommyTalks in action.](Ui.png)
 
 Ever wondered if there was an easier way to keep track of your tasks? :thinking:\
 Look no further, introducing **Tommy** - _your one-stop task manager_\
@@ -8,8 +8,8 @@ Look no further, introducing **Tommy** - _your one-stop task manager_\
 track of their completion status all in one place!
 ---
 
-## Adding Tasks
-TommyTalks supports 3 kinds of tasks:
+## Adding tasks
+tommyTalks supports 3 kinds of tasks:
 
 - **ToDo**
 - **Deadline**
@@ -222,12 +222,12 @@ Bye, hope to see you again!
 ---
 ## Summary
 
-For a quick reference, TommyTalks supports 3 kinds of tasks:
+For a quick reference, tommyTalks supports 3 kinds of tasks:
 1. [ToDo](#todo)
 2. [Deadline](#deadline)
 3. [Event](#event)
 
-TommyTalks has the following commands you can call:
+tommyTalks has the following commands you can call:
 1. [List](#listing-all-tasks-list)
 2. [Mark](#mark-as-completed-mark)
 3. [Unmark](#unmark-a-task-unmark)
