+++ src/main/java/commands/Command.java
@@ -1,7 +1,7 @@
-package Commands;
+package commands;
 
-import TommyTalks.Storage;
-import TommyTalks.Ui;
+import tommyTalks.Storage;
+import tommyTalks.Ui;
 
 /**
  * Parent class for commands that user inputs.
