+++ src/main/java/tommyTalks/TommyTalks.java
@@ -1,4 +1,4 @@
-package TommyTalks;
+package tommyTalks;
 
 import javafx.application.Application;
